api2/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
pytz